Refactor bandwidth limiter tests into focused modules
