+++ crates/bandwidth/src/limiter/tests/configuration.rs
@@ -1,0 +1,104 @@
+use super::{BandwidthLimiter, LimiterChange, MINIMUM_SLEEP_MICROS, recorded_sleep_session};
+use std::num::NonZeroU64;
+
+#[test]
+fn limiter_update_limit_resets_internal_state() {
+    let mut session = recorded_sleep_session();
+    session.clear();
+
+    let new_limit = NonZeroU64::new(8 * 1024 * 1024).unwrap();
+    let mut baseline = BandwidthLimiter::new(new_limit);
+    baseline.register(4096);
+    let baseline_sleeps = session.take();
+
+    session.clear();
+
+    let mut limiter = BandwidthLimiter::new(NonZeroU64::new(1024).unwrap());
+    limiter.register(4096);
+    session.clear();
+
+    limiter.update_limit(new_limit);
+    limiter.register(4096);
+    assert_eq!(limiter.limit_bytes(), new_limit);
+    assert_eq!(limiter.recommended_read_size(1 << 20), 1 << 20);
+
+    let updated_sleeps = session.take();
+    assert_eq!(updated_sleeps, baseline_sleeps);
+}
+
+#[test]
+fn limiter_update_configuration_resets_state_and_updates_burst() {
+    let mut session = recorded_sleep_session();
+    session.clear();
+
+    let initial_limit = NonZeroU64::new(1024).unwrap();
+    let initial_burst = NonZeroU64::new(4096).unwrap();
+    let mut limiter = BandwidthLimiter::with_burst(initial_limit, Some(initial_burst));
+    limiter.register(8192);
+    assert!(limiter.accumulated_debt_for_testing() > 0);
+
+    let new_limit = NonZeroU64::new(8 * 1024 * 1024).unwrap();
+    let new_burst = NonZeroU64::new(2048).unwrap();
+    limiter.update_configuration(new_limit, Some(new_burst));
+
+    assert_eq!(limiter.limit_bytes(), new_limit);
+    assert_eq!(limiter.burst_bytes(), Some(new_burst));
+    assert_eq!(limiter.accumulated_debt_for_testing(), 0);
+
+    session.clear();
+    limiter.register(1024);
+    let recorded = session.take();
+    assert!(
+        recorded.is_empty()
+            || recorded
+                .iter()
+                .all(|duration| duration.as_micros() <= MINIMUM_SLEEP_MICROS)
+    );
+}
+
+#[test]
+fn limiter_reset_clears_state_and_preserves_configuration() {
+    let mut session = recorded_sleep_session();
+    session.clear();
+
+    let limit = NonZeroU64::new(1024).unwrap();
+    let mut baseline = BandwidthLimiter::new(limit);
+    baseline.register(4096);
+    let baseline_sleeps = session.take();
+
+    session.clear();
+
+    let mut limiter = BandwidthLimiter::new(limit);
+    limiter.register(4096);
+    assert!(limiter.accumulated_debt_for_testing() > 0);
+
+    session.clear();
+
+    limiter.reset();
+    assert_eq!(limiter.limit_bytes(), limit);
+    assert_eq!(limiter.burst_bytes(), None);
+    assert_eq!(limiter.accumulated_debt_for_testing(), 0);
+
+    limiter.register(4096);
+    let reset_sleeps = session.take();
+    assert_eq!(reset_sleeps, baseline_sleeps);
+}
+
+#[test]
+fn limiter_change_helper_methods_reflect_state() {
+    assert!(!LimiterChange::Unchanged.is_changed());
+    assert!(!LimiterChange::Unchanged.leaves_limiter_active());
+    assert!(!LimiterChange::Unchanged.disables_limiter());
+
+    assert!(LimiterChange::Enabled.is_changed());
+    assert!(LimiterChange::Enabled.leaves_limiter_active());
+    assert!(!LimiterChange::Enabled.disables_limiter());
+
+    assert!(LimiterChange::Updated.is_changed());
+    assert!(LimiterChange::Updated.leaves_limiter_active());
+    assert!(!LimiterChange::Updated.disables_limiter());
+
+    assert!(LimiterChange::Disabled.is_changed());
+    assert!(!LimiterChange::Disabled.leaves_limiter_active());
+    assert!(LimiterChange::Disabled.disables_limiter());
+}
