+++ crates/bandwidth/src/limiter/tests/mod.rs
@@ -1,0 +1,11 @@
+pub(super) use super::{
+    BandwidthLimiter, LimiterChange, MAX_SLEEP_DURATION, MINIMUM_SLEEP_MICROS,
+    RecordedSleepSession, apply_effective_limit, duration_from_microseconds,
+    recorded_sleep_session, recorded_sleeps, sleep_for,
+};
+
+mod apply_effective_limit_cases;
+mod configuration;
+mod helpers;
+mod pacing;
+mod recording;
